Don't sample every pixel of large logos
The analyze_img code examined every pixel of the image; for large
images, this caused the page to time out, thus making it impossible to
change the image away from the problematically-large one.

Subsample the image if it is more than 200px on a side, by sampling in
an even 200x200 grid.  200px was chosen as an arbitrary upper limit; 40k
points is still a large number to sample, but in testing can be done in
less than 10s.

Fixes #29929.